requirements.txt
numpy==1.24.3
onnxruntime==1.15.1
insightface==0.7.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pandas==2.1.3
python-dotenv==1.0.0
supabase==2.0.3
opencv-python==4.8.1.78
